refact: remove redundant / stale discovery queries/state
