+++ src/js/features/search/hooks.ts
@@ -25,3 +25,15 @@ export const useNonFilterQueryParams = (): QueryParams => {
     return qp;
   }, [mode, textQuery, textQueryStatus]);
 };
+
+export const useSearchableFields = () => {
+  /**
+   * Hook which calculates a set of searchable fields (which share IDs with charts), which can be used, for example, to
+   * choose whether to add a click event to a chart for the field.
+   */
+  const { filterSections } = useSearchQuery();
+  return useMemo(
+    () => new Set(filterSections.flatMap((section) => section.fields).map((field) => field.id)),
+    [filterSections]
+  );
+};
